+++ example/main.nim
@@ -1,25 +1,38 @@
 #[
   Created at: 07/04/2021 11:12:25 Sunday
-  Modified at: 07/04/2021 01:21:14 PM Sunday
+  Modified at: 07/04/2021 11:51:53 PM Sunday
 ]#
+
+{.experimental: "codeReordering".}
 
 import jsconsole
 import gm_api
+import asyncjs
+import jsffi
+
 
 when isMainModule:
   const metadataBlock = genMetadataBlock(
-    name = "Userscript example\nconsole.log('test')",
+    name = "Userscript example",
     nameTranslations = {"pt-br": "Exemplo de userscript"},
     description = "A simple example to test the GM bindings in Nim",
-    descriptionTranslations = {"pt-br": "Um simples exemplo para testar os bindings do GM no Nim"},
+    descriptionTranslations = {"pt-br": "Um simples exemplo para testar bindings para o GM no Nim"},
     grant = [
-      GmPermitions.info
+      GmPermitions.info,
+      GmPermitions.listValues,
+      GmPermitions.setValue,
+      GmPermitions.getValue,
+      GmPermitions.notification,
+      GmPermitions.openInTab,
+      GmPermitions.registerMenuCommand,
+      GmPermitions.setClipboard,
+      GmPermitions.xmlHttpRequest,
     ],
     require = [
-      # "http://127.0.0.1/u3/apache/www/libs/button/button.js",
+      # "https://arantius.com/misc/greasemonkey/imports/greasemonkey4-polyfill.js",
     ],
     resource = {
-      "customCss": "http://127.0.0.1/u3/apache/www/libs/button/button.css"
+      "avatar": "https://gitea.com/avatars/be2da4e8a9cbffdc433db9058105a1ce?size=112"
     },
     homepageUrl = "http://localhost/home",
     supportUrl = "http://localhost/support",
@@ -29,11 +42,9 @@ when isMainModule:
     icon = "https://gitea.com/avatars/be2da4e8a9cbffdc433db9058105a1ce?size=112",
     match = [
       "*://localhost/*",
-      "*://127.0.0.1/*",
     ],
     `include` = [
-      "*://localhost/*",
-      "*://localhost/",
+      "*://127.0.0.1/*",
     ],
     excludeMatch = [
       "*://duckduckgo/*",
@@ -49,4 +60,35 @@ when isMainModule:
   {.emit: metadataBlock.} # Use the metadatablock into js code
 
   console.log metadataBlock
-  console.log(GM.info())
+  discard main()
+
+proc main {.async.} =
+  # console.log(GM.info()) # Not working
+  console.log(await GM.listValues())
+  console.log(await GM.getValue("a", "unknown"))
+  console.log(unsafeWindow)
+  GM.registerMenuCommand("testMenu", (proc() =
+    console.log("menu clicked")),
+  "t")
+  GM.registerMenuCommand("copy 'Nim' to clipboard", (proc() =
+    GM.setClipboard("Nim")
+    console.log "copied"),
+  "t")
+  # console.log GM.xmlHttpRequest()
+  GM.notification("text", "title", onclick = (proc() =
+    console.log("CLICK")
+    GM.openInTab("http://localhost", false)),
+  ondone = (proc() =
+    console.log("DONE")))
+
+  GM.notification("second", "title", onclick = (proc() =
+    console.log("CLICK")))
+
+  await GM.setValue("a", "test")
+  console.log(await GM.listValues())
+
+  GM.xmlHttpRequest("http://localhost/", "GET", onload = (proc(a: JsObject) =
+      console.log a),
+    headers = {
+      "referrer": "about:blank"
+    })