+++ src/gm_api.nim
@@ -1,7 +1,18 @@
 #[
   Created at: 07/04/2021 10:58:31 Sunday
-  Modified at: 07/04/2021 01:23:49 PM Sunday
+  Modified at: 07/04/2021 11:50:45 PM Sunday
 ]#
+
+##[
+  Reference for violentmonkey: https://violentmonkey.github.io/api/gm/#gm_info
+  Reference for greasymonkey api: https://wiki.greasespot.net/Greasemonkey_Manual:API
+]##
+
+import std/asyncjs
+from std/dom import nil
+from std/jsffi import JsObject, `[]=`, newJsObject
+
+import jsconsole
 
 type
   GmType = object
@@ -28,130 +39,109 @@ type
     browserVersion*: cstring
     os*: GmInfoOs
 
-proc info*(gm: GmType): GmInfo {.importcpp.}
+type Callable = concept x
+  x()
 
+# Gm object export
 let GM* {.importc, noDecl.}: GmType
 
+# bindings
+{.push importcpp.}
+proc info*(gm: GmType): GmInfo
+  ## NOT WORKING
+  ## Exposes this information (plus a bit more) to the user script.
+
+# values
+proc getValue*(gm: GmType; key,
+               defaultValue: cstring): Future[cstring] {.async.}
+  ## Retrieves stored values, see GM.setValue below.
+proc setValue*(gm: GmType; key, value: cstring) {.async.}
+  ## Permanently stores a value under a key, later available via GM.getValue.
+proc deleteValue*(gm: GmType; key: cstring) {.async.}
+  ## Deletes a value from chrome that was previously set.
+proc listValues*(gm: GmType): Future[seq[cstring]] {.async.}
+  ## Retrieves an array of stored values' keys.
+
+# # other
+proc notification(gm: GmType; data: JsObject)
+proc notification(gm: GmType; text: cstring; title, image = "".cstring;
+                   onclick: Callable = proc())
+  ## Opens a notification dialog.
+
+proc openInTab*(gm: GmType; url: cstring; openInBackground = false)
+  ## Opens a given URL in a new tab.
+proc registerMenuCommand*(gm: GmType; caption: cstring; commandFunc: Callable;
+                          accessKey: cstring)
+  ## Adds an item to the "User Script Commands" section of the Monkey Menu.
+proc setClipboard*(gm: GmType; text: cstring)
+  ## Sets the contents of the clipboard.
+proc xmlHttpRequest(gm: GmType; details: JsObject)
+
+# proc addValueChangeListener*[T](
+#   gm: GmType; name: cstring;
+#   callback: proc(name, oldValue, newValue: cstring; remote: bool): T
+# ) #[ {.hint: "Avaliable just for Violentmonkey".} ]#
+{.pop.}
+
+proc notification_placeholderCallback() = discard
+proc notification*(gm: GmType; text: cstring; title, image = "".cstring;
+                   onclick, ondone: Callable = notification_placeholderCallback) =
+  ## Opens a notification dialog.
+  var jsObj = newJsObject()
+  jsObj["text"] = text
+  jsObj["title"] = title
+  jsObj["image"] = image
+  jsObj["onclick"] = onclick
+  jsObj["ondone"] = ondone
+  GM.notification(jsObj)
+
+type GmHttpResponse = concept x
+  x(newJsObject())
+
+proc push(obj: JsObject, val: JsObject) {.importcpp.}
+
+proc xmlHttpRequest_placeholderCallback(x: JsObject) = discard
+proc xmlHttpRequest*(gm: GmType; url, `method`: cstring;
+                     binary = false; context = newJsObject(); data = "".cstring;
+                     headers: openArray[(string, string)] = [],
+                     overrideMimeType, user, password = "".cstring;
+                     synchronous = false; timeout = 0; upload = newJsObject();
+                     onabort, onerror, onload, onprogress, onreadystatechange,
+                     ontimeout: GmHttpResponse = xmlHttpRequest_placeholderCallback) =
+  ## A variant of XMLHttpRequest, this method allows skipping use the same-origin policy, enabling complex mashups.
+  var jsObj = newJsObject()
+  jsObj["url"] = url
+  jsObj["method"] = `method`
+  jsObj["binary"] = binary
+  jsObj["context"] = context
+  jsObj["data"] = data
+
+  var jsHeaders = newJsObject()
+  for i, (key, val) in headers:
+    jsHeaders[key.cstring] = val.cstring
+
+  jsObj["headers"] = jsHeaders
+
+
+  jsObj["overrideMimeType"] = overrideMimeType
+  jsObj["user"] = user
+  jsObj["password"] = password
+
+
+  # events
+  jsObj["onabort"] = onabort
+  jsObj["onerror"] = onerror
+  jsObj["onload"] = onload
+  jsObj["onprogress"] = onprogress
+  jsObj["onreadystatechange"] = onreadystatechange
+  jsObj["ontimeout"] = ontimeout
+  console.log(jsObj)
+  GM.xmlHttpRequest(jsObj)
+
+var unsafeWindow* {.importc, noDecl.}: dom.Window
+  ## This object provides access to the raw JavaScript window scope of the content page. It is most commonly used to access JavaScript variables on the page.
+# hint: "This command can open certain security holes in your user script, and it is recommended to use this command sparingly."
+
 # Metadata block
-
-type
-  GmPermitions* {.pure.} = enum
-    info, getValue, setValue, deleteValue, listValues, addValueChangeListener,
-      removeValueChangeListener, getResourceText, getResourceURL, addStyle,
-      openInTab, registerMenuCommand, unregisterMenuCommand, notification,
-      setClipboard, xmlhttpRequest, download, winClose, winFocus
-
-  GmRunAt* {.pure.} = enum
-    ## Decide when the script will execute.
-    docStart,
-    docBody,
-    docEnd,
-    docIdle
-  GmInjectInto* {.pure.} = enum
-    ## Decide which context the script will be injected into.
-    page, content, `auto`
-
-proc `$`*(self: GmPermitions): string =
-  case self:
-  of GmPermitions.info: "GM.info"
-  of GmPermitions.getValue: "GM.getValue"
-  of GmPermitions.setValue: "GM.setValue"
-  of GmPermitions.deleteValue: "GM.deleteValue"
-  of GmPermitions.listValues: "GM.listValues"
-  of GmPermitions.addValueChangeListener: "GM.addValueChangeListener"
-  of GmPermitions.removeValueChangeListener: "GM.removeValueChangeListener"
-  of GmPermitions.getResourceText: "GM.getResourceText"
-  of GmPermitions.getResourceURL: "GM.getResourceURL"
-  of GmPermitions.addStyle: "GM.addStyle"
-  of GmPermitions.openInTab: "GM.openInTab"
-  of GmPermitions.registerMenuCommand: "GM.registerMenuCommand"
-  of GmPermitions.unregisterMenuCommand: "GM.unregisterMenuCommand"
-  of GmPermitions.notification: "GM.notification"
-  of GmPermitions.setClipboard: "GM.setClipboard"
-  of GmPermitions.xmlhttpRequest: "GM.xmlhttpRequest"
-  of GmPermitions.download: "GM.download"
-  of GmPermitions.winClose: "window.close"
-  of GmPermitions.winFocus: "window.focus"
-
-proc `$`*(self: GmRunAt): string =
-  case self:
-  of GmRunAt.docStart: "document-start"
-  of GmRunAt.docBody: "document-body"
-  of GmRunAt.docEnd: "document-end"
-  of GmRunAt.docIdle: "document-idle"
-
-from strformat import fmt
-from strutils import join, unindent, replace
-
-proc genMetadataBlock*(
-  # project
-  name: string;                                              ## Required *
-  nameTranslations: openArray[(string, string)] = [];        ## Optional
-  namespace = "Violentmonkey Scripts";                       ## Optional
-  description = "New Nim userscript";                        ## Optional
-  descriptionTranslations: openArray[(string, string)] = []; ## Optional
-  version = "1.0.0";                                         ## Optional
-  author = "You";                                            ## Optional
-  icon = "";                                                 ## Optional
-  require: openArray[string] = [];                           ## Optional
-  resource: openArray[(string, string)] = [];                ## Optional
-  supportUrl = "";                                           ## Optional
-  homepageUrl = "";                                          ## Optional
-
-
-  # matching
-  match: openArray[string] = ["*://*/*"]; # # Default matches everything
-  excludeMatch: openArray[string] = [];                      ## Optional
-  `include`: openArray[string] = [];       ## Optional (prefers `match`)
-  exclude: openArray[string] = [];  ## Optional (prefers `excludematch`)
-
-  # Advanced
-  grant: openArray[GmPermitions] = []; ## Optional
-  runAt = GmRunAt.docIdle;            ## Optional
-  downloadUrl = "";                   ## Optional
-  noFrames = false
-): string {.compileTime.} =
-  ## Generates the userscript metadata block and returns it
-  var metadatas = newSeq[string]()
-
-  # helpers
-  proc newMetadata(key, value: string): string =
-    fmt"// @{key} " & value.replace("\n", "\\n")
-  proc addMeta(key, value: string) =
-    if value.len > 0: metadatas.add newMetadata(key, value)
-
-  # adding
-
-  # project
-  addMeta("name", name)
-  for (key, name) in nameTranslations: addMeta(fmt"name:{key}", name)
-  addMeta("namespace", namespace)
-  addMeta("description", description)
-  for (key, desc) in descriptionTranslations: addMeta(fmt"description:{key}", desc)
-  addMeta("version", version)
-  addMeta("author", author)
-  addMeta("icon", icon)
-  for req in require: addMeta(fmt"require", req)
-  for (key, res) in resource: addMeta("resource", fmt"{key} {res}")
-  addMeta("supportURL", supportUrl)
-  addMeta("homepageURL", homepageUrl)
-
-  # matching
-  for mat in match: addMeta("match", mat)
-  for exclMat in excludeMatch: addMeta("exclude-match", exclMat)
-  for incl in `include`: addMeta("include", incl)
-  for excl in exclude: addMeta("exclude", excl)
-
-  # advanced
-  if grant.len > 0:
-    for perm in grant: addMeta("grant", $perm)
-  else: addMeta("grant", "none")
-  addMeta("run-at", $runAt)
-  addMeta("downloadURL", downloadUrl)
-  if noFrames: addMeta("noframes", "")
-
-  # return
-  result = fmt"""// ==UserScript==
-               {metadatas.join("\n")}
-               // ==/UserScript==
-               // Metadata automatically generated""".unindent
+import ./gm_api/metadata
+export metadata